+++ src/api/config.ts
@@ -1,0 +1,1 @@
+export const API_BASE_URL: string = import.meta.env.VITE_API_BASE_URL
