+++ src/api/notification/notification.js
@@ -1,4 +1,4 @@
-import { authenticatedFetch } from '../auth/authUtils'
+import { authenticatedFetch } from '../authUtils'
 import { API_BASE_URL } from '../config'
 
 const getNotification = async lastNotificationId => {
