+++ src/api/pictures/deletePicture.js
@@ -1,4 +1,4 @@
-import { authenticatedFetch } from '../auth/authUtils'
+import { authenticatedFetch } from '../authUtils'
 import { API_BASE_URL } from '../config'
 
 const deleteAlbumPicture = async (albumId, selectedPictureIds) => {
